precompiles: Expose internal API
Expose all `xxx_analyze()` and `xxx_execute()` functions from
the `evmone::state` library. They are usable for tests and benchmarks.

Make `EVMONE_PRECOMPILES_SILKPRE` compile definition available to
the `evmone::state`s users.
